# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AssetFlow Enterprise Asset & Resource Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List assets with optional filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: integer
        - name: department
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Register a new asset (asset_manager, admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get asset detail with allocation and maintenance history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetDetail"
    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Update asset information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /allocations:
    get:
      operationId: listAllocations
      tags: [allocations]
      summary: List all allocations
      parameters:
        - name: assetId
          in: query
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of allocations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AllocationWithDetails"
    post:
      operationId: createAllocation
      tags: [allocations]
      summary: Allocate an asset to an employee (409 if not available)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllocationInput"
      responses:
        "201":
          description: Allocated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allocation"
        "409":
          description: Asset not available (includes current holder info)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllocationConflict"
...
  /allocations/{id}/return:
    post:
      operationId: returnAllocation
      tags: [allocations]
      summary: Return an allocated asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnInput"
      responses:
        "200":
          description: Returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allocation"
...
components:
  schemas:
...
    Asset:
      type: object
      required: [id, assetTag, name, status, createdAt]
      properties:
        id:
          type: integer
        assetTag:
          type: string
        name:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        acquisitionDate:
          type: ["string", "null"]
        acquisitionCost:
          type: ["number", "null"]
        condition:
          type: string
        location:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        isBookable:
          type: boolean
        status:
          type: string
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    AssetDetail:
      type: object
      required: [id, assetTag, name, status, createdAt, allocationHistory, maintenanceHistory]
      properties:
        id:
          type: integer
        assetTag:
          type: string
        name:
          type: string
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        acquisitionDate:
          type: ["string", "null"]
        acquisitionCost:
          type: ["number", "null"]
        condition:
          type: string
        location:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        isBookable:
          type: boolean
        status:
          type: string
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        allocationHistory:
          type: array
          items:
            $ref: "#/components/schemas/AllocationWithDetails"
        maintenanceHistory:
          type: array
          items:
            $ref: "#/components/schemas/MaintenanceRequest"
    AssetInput:
      type: object
      required: [assetTag, name]
      properties:
        assetTag:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        categoryId:
          type: ["integer", "null"]
        serialNumber:
          type: ["string", "null"]
        acquisitionDate:
          type: ["string", "null"]
        acquisitionCost:
          type: ["number", "null"]
        condition:
          type: string
        location:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        isBookable:
          type: boolean
        status:
          type: string
        departmentId:
          type: ["integer", "null"]
    AssetUpdate:
      type: object
      properties:
        name:
          type: string
        categoryId:
          type: ["integer", "null"]
        serialNumber:
          type: ["string", "null"]
        acquisitionDate:
          type: ["string", "null"]
        acquisitionCost:
          type: ["number", "null"]
        condition:
          type: string
        location:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        isBookable:
          type: boolean
        status:
          type: string
        departmentId:
          type: ["integer", "null"]
    Allocation:
      type: object
      required: [id, assetId, employeeId, allocatedAt, status]
      properties:
        id:
          type: integer
        assetId:
          type: integer
        employeeId:
          type: integer
        departmentId:
          type: ["integer", "null"]
        allocatedAt:
          type: string
        expectedReturnDate:
          type: ["string", "null"]
        returnedAt:
          type: ["string", "null"]
        returnConditionNotes:
          type: ["string", "null"]
        status:
          type: string
        createdBy:
          type: ["integer", "null"]
    AllocationWithDetails:
      type: object
      required: [id, assetId, employeeId, allocatedAt, status, isOverdue]
      properties:
        id:
          type: integer
        assetId:
          type: integer
        assetTag:
          type: ["string", "null"]
        assetName:
          type: ["string", "null"]
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        allocatedAt:
          type: string
        expectedReturnDate:
          type: ["string", "null"]
        returnedAt:
          type: ["string", "null"]
        returnConditionNotes:
          type: ["string", "null"]
        status:
          type: string
        isOverdue:
          type: boolean
    AllocationInput:
      type: object
      required: [assetId, employeeId]
      properties:
        assetId:
          type: integer
        employeeId:
          type: integer
        departmentId:
          type: ["integer", "null"]
        expectedReturnDate:
          type: ["string", "null"]
    AllocationConflict:
      type: object
      required: [error, currentHolderName, allocationId]
      properties:
        error:
          type: string
        currentHolderName:
          type: string
        currentHolderDepartment:
          type: ["string", "null"]
        allocationId:
          type: integer
    ReturnInput:
      type: object
      properties:
        returnConditionNotes:
          type: ["string", "null"]
...
    MaintenanceRequest:
      type: object
      required: [id, assetId, raisedBy, issueDescription, priority, status, createdAt]
      properties:
        id:
          type: integer
        assetId:
          type: integer
        assetName:
          type: ["string", "null"]
        assetTag:
          type: ["string", "null"]
        raisedBy:
          type: integer
        raisedByName:
          type: ["string", "null"]
        issueDescription:
          type: string
        priority:
          type: string
        photoUrl:
          type: ["string", "null"]
        status:
          type: string
        assignedTechnician:
          type: ["string", "null"]
        approvedBy:
          type: ["integer", "null"]
        resolvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...